tests/wpt/web-platform-tests/notifications/Notification-permission.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: Notification - permission (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="4.3 Permission" href="http://www.w3.org/TR/notifications/#permission">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
    test(function() {
      assert_in_array(Notification.permission, ["default", "denied", "granted"], "Notification.permission");
    }, "Test checks that notification the value of permission is in array");
  </script>
</body>
